Fix Java idle-timeout heartbeat delay overflowing int above ~24.8 days (#5617)
(cherry picked from commit c718bdccfe201c20294f726dc1c1d0b2246d4da9)

+++ java/src/com.zeroc.ice/src/main/java/com/zeroc/Ice/IdleTimeoutTransceiverDecorator.java
@@ -170,6 +170,6 @@ final class IdleTimeoutTransceiverDecorator implements Transceiver {
     private void rescheduleWriteTimer() {
         cancelWriteTimer();
         _writeTimerFuture =
-            _scheduledExecutorService.schedule(_sendHeartbeat, _idleTimeout * 1000 / 2, TimeUnit.MILLISECONDS);
+            _scheduledExecutorService.schedule(_sendHeartbeat, (long) _idleTimeout * 1000 / 2, TimeUnit.MILLISECONDS);
     }
 }
